Move ExtensionTableXXX
